pyproject.toml
[project]
name = "voiceauth"
version = "1.0.0"
description = "Voice Authentication Server"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.7.0",
    "sherpa-onnx>=1.10.0",
    "numpy>=1.26.0",
    "av>=14.0.0",
    "ulid-py>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["voiceauth"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pyright>=1.1.408",
    "ruff>=0.9.6",
    "coverage>=7.6.10",
    "pre-commit>=4.1.0",
    "pytest-cov>=7.0.0",
    "alembic>=1.14.0",
]

[tool.pyright]
typeCheckingMode = "basic"
exclude = ["venv", ".venv", "alembic", "packages"]
reportMissingTypeStubs = false

[tool.ruff]
target-version = "py311"
exclude = ["alembic", "packages"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
]
extend-safe-fixes = ["F401"]  # Allow auto-removal of unused imports

[tool.pytest.ini_options]
testpaths = ["voiceauth"]
pythonpath = ["."]
addopts = ["--import-mode=importlib"]

[tool.coverage.run]
source = ["voiceauth"]

[tool.coverage.report]
show_missing = true
